Add files via upload
Property taxes account for a large portion of revenue for local governments, allowing them to
fund police, schools, pension liabilities and many other services provided to citizens. In the
United States, property taxes are collected at the county level, requiring owners to pay a sum
that is based on tax rates, along with levies for their locale, multiplied by the estimated or
assessed value of the property. It is the role of a government entity, commonly knowns as the
Assessor, to determine the fair market value of each property.
For this project we use the example of the Cook County Assessor’s Office (CCAO), which is
responsible for the second most populous county in the United States, and home to the City of
Chicago and over 130 other municipalities. To get a sense of the magnitude of the task for this
organization, in FY 2019, CCAO was in charge of valuing over 1.8 million properties,
totaling$15.58 billion in taxes.
Determining what is a fair market value with this diversity of parcels that are spread across over
130 municipalities is one of the most complex valuation tasks in the United States. Despite its
complexity, prior to 2018, CCAO’s valuation process was obscure and inaccurate. Compounded
with unusually high tax rates and the flawed modeling process, CCAO, which is overseen by an
elected official, faced intense public and political pressures. One outcome was the surprising, yet
decisive win in the 2018 election of Fritz Kaegi, who promised to reform CCOA and make it
transparent.
To address this challenge, the new Assessor invested in technology, has increased the
transparency of data and has formed and later expanded the data science team in the agency.
The modeling techniques underlying the valuation pipeline have also shifted from old
underperforming technology using SPSS, to a more capable implementation in R. With respect
to transparency, CCAO is making all their choices, including modeling and data decisions,
publicly available.
For additional background and context on the data science efforts at CCAO, review the
documentation the data science team at CCAO generated, especially for their machine learning
models. Additionally, view the video recording of the presentation CCOA’s data scientist Dan
Snow gave last year about their machine learning modeling and data.